+++ db/.subgraph-modules.env
@@ -1,0 +1,8 @@
+# get actual root dir of repo.
+# ~/gh/cncf/landscape-graph/docs/resources ᐅ git rev-parse --show-toplevel  
+# /home/myoung/gh/cncf/landscape-graph
+
+export LANDSCAPE_GIT_REPO_ROOT=$(git rev-parse --show-toplevel)
+export LANDGRAPH_CLI=$LANDSCAPE_GIT_REPO_ROOT/util/landgraph/landgraph.sh
+
+alias landgraph=$LANDGRAPH_CLI
