+++ db/cncf/x-cncf-member-schema.graphql
@@ -1,0 +1,127 @@
+type Acquisi0tions {
+  date: String
+  acquiree: String
+}
+
+type CrunchbaseData {
+  name: String
+  description: String
+  homepage: String
+  city: String
+  region: String
+  country: String
+  twitter: String
+  linkedin: String
+  ticker: String
+  funding: Int
+  stockExchange: String
+  company_type: String
+  kind: String
+  numEmployeesMin: Int
+  numEmployeesMax: Int
+  industries: [String]
+  parents: [String]
+  acquisitions: [Acquisitions]
+
+}
+
+type LatestTweetDate {
+  text: String
+  value: String
+  original: String
+}
+
+type LatestCommitDate {
+  text: String
+  value: String
+  original: String
+}
+
+type ImageData {
+  fileName: String
+  hash: String
+}
+
+type YahooFinanceData {
+  market_cap: Int
+  effective_ticker: String
+}
+
+type GithubStartCommitData {
+  start_commit_link: String
+  start_date: String
+}
+
+type Repos {
+  url: String
+  stars: Int
+}
+
+type Languages {
+  name: String
+  value: Int
+  color: String
+}
+
+type GithubData {
+  contributions: String
+  firstWeek: String
+  stars: Int
+  license: String
+  description: String
+  latest_commit_date: String
+  latest_commit_link: String
+  contributors_count: Int
+  contributors_link: String
+  languages: [Languages]
+}
+
+type AutogeneratedMainType {
+  name: String
+  homepage_url: String
+  repo_url: String
+  logo: String
+  twitter: String
+  crunchbase: String
+  firstCommitDate: String
+  firstCommitLink: String
+  latestCommitLink: String
+  commitsThisYear: Int
+  contributorsCount: Int
+  contributorsLink: String
+  language: String
+  stars: Int
+  license: String
+  headquarters: String
+  description: String
+  organization: String
+  path: String
+  landscape: String
+  category: String
+  amountKind: String
+  amount: Int
+  oss: Boolean
+  href: String
+  bestPracticeBadgeId: Int
+  bestPracticePercentage: Int
+  starsPresent: Boolean
+  starsAsText: String
+  marketCapPresent: Boolean
+  marketCapAsText: String
+  id: String
+  flatName: String
+  member: String
+  isSubsidiaryProject: Boolean
+  industries: [String]
+  crunchbaseData: CrunchbaseData
+  latestTweetDate: LatestTweetDate
+  latestCommitDate: LatestCommitDate
+  image_data: ImageData
+  yahoo_finance_data: YahooFinanceData
+  github_start_commit_data: GithubStartCommitData
+  repos: [Repos]
+  github_data: GithubData
+}
+
+# Types with identical fields:
+# LatestTweetDate LatestCommitDate
