+++ db/cncf/x-cncf-project-schema.graphql
@@ -1,0 +1,139 @@
+type Acquisitions {
+  date: String
+  acquiree: String
+}
+
+type CrunchbaseData {
+  name: String
+  description: String
+  homepage: String
+  city: String
+  region: String
+  country: String
+  twitter: String
+  linkedin: String
+  funding: Int
+  company_type: String
+  kind: String
+  numEmployeesMin: Int
+  numEmployeesMax: Int
+  industries: [String]
+  parents: [String]
+  acquisitions: [Acquisitions]
+}
+
+type LatestTweetDate {
+  text: String
+  value: String
+  original: String
+}
+
+type ReleaseDate {
+  text: String
+  value: String
+  original: String
+}
+
+type LatestCommitDate {
+  text: String
+  value: String
+  original: String
+}
+
+type ImageData {
+  fileName: String
+  hash: String
+}
+
+type GithubStartCommitData {
+  start_commit_link: String
+  start_date: String
+}
+
+type Repos {
+  url: String
+  stars: Int
+}
+
+type Languages {
+  name: String
+  value: Int
+  color: String
+}
+
+type GithubData {
+  contributions: String
+  firstWeek: String
+  stars: Int
+  license: String
+  description: String
+  latest_commit_date: String
+  latest_commit_link: String
+  release_date: String
+  release_link: String
+  contributors_count: Int
+  contributors_link: String
+  languages: [Languages]
+}
+
+type Extra {
+  accepted: String
+  dev_stats_url: String
+  blog_url: String
+  artwork_url: String
+  slack_url: String
+  chat_channel: String
+}
+
+type AutogeneratedMainType {
+  name: String
+  homepage_url: String
+  project: String
+  repo_url: String
+  logo: String
+  twitter: String
+  crunchbase: String
+  firstCommitDate: String
+  firstCommitLink: String
+  latestCommitLink: String
+  releaseLink: String
+  commitsThisYear: Int
+  contributorsCount: Int
+  contributorsLink: String
+  language: String
+  stars: Int
+  license: String
+  headquarters: String
+  description: String
+  organization: String
+  path: String
+  landscape: String
+  category: String
+  amountKind: String
+  amount: Int
+  oss: Boolean
+  href: String
+  bestPracticeBadgeId: Int
+  bestPracticePercentage: Int
+  starsPresent: Boolean
+  starsAsText: String
+  marketCapPresent: Boolean
+  marketCapAsText: String
+  id: String
+  flatName: String
+  relation: String
+  isSubsidiaryProject: Boolean
+  industries: [String]
+  crunchbaseData: CrunchbaseData
+  latestTweetDate: LatestTweetDate
+  releaseDate: ReleaseDate
+  latestCommitDate: LatestCommitDate
+  image_data: ImageData
+  github_start_commit_data: GithubStartCommitData
+  repos: [Repos]
+  github_data: GithubData
+  extra: Extra
+}
+
+# Types with identical fields:
+# LatestTweetDate ReleaseDate LatestCommitDate
