--- util/gen-exec-schema/gen-exec-schema.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-
-# TODO moar
-node index.js --sdl ../../db/cncf/cncf.graphql -o ../../db/cncf/cncf.full.graphql -v -gserver
