+++ util/landgraph/.dockerignore
@@ -1,0 +1,7 @@
+.dockerignore
+node_modules
+npm-debug.log
+Dockerfile
+.git
+.gitignore
+#.npmrc
