+++ util/landgraph/.env.example
@@ -1,0 +1,3 @@
+NEO4J_USER=neo4j
+NEO4J_PASSWORD=landscape
+NEO4J_URI=bolt://localhost:7687
