+++ util/landgraph/Dockerfile
@@ -1,0 +1,22 @@
+#
+# https://snyk.io/blog/10-best-practices-to-containerize-nodejs-web-applications-with-docker
+# 
+# 
+# --------------> The build image
+FROM node:latest AS build
+WORKDIR /usr/src/app
+COPY package*.json /usr/src/app/
+#RUN --mount=type=secret,mode=0644,id=npmrc,target=/usr/src/app/.npmrc npm ci --only=production
+RUN npm ci --only=production
+
+# --------------> The production image
+FROM node:lts-alpine@sha256:2c405ed42fc0fd6aacbe5730042640450e5ec030bada7617beac88f742b6997b
+RUN apk add dumb-init
+ENV NODE_ENV production
+USER node
+WORKDIR /usr/src/app
+COPY --chown=node:node --from=build /usr/src/app/node_modules /usr/src/app/node_modules
+COPY --chown=node:node . /usr/src/app
+CMD ["dumb-init", "node", "gen-exec-schema.js"]
+
+
