+++ util/landgraph/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@cncfgraph/gen-exec-schema",
+  "name": "@cncfgraph/landgraph",
   "version": "1.0.0",
   "description": "Generate an executable schema from GraphQL type definitions",
-  "main": "gen-exec-schema.js",
+  "main": "landgraph.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -14,6 +14,7 @@
     "apollo-server": "^3.10.1",
     "command-line-args": "^5.2.1",
     "command-line-usage": "^6.1.3",
+    "dockly": "^3.22.2",
     "dotenv": "^16.0.1",
     "graphql": "^16.5.0",
     "neo4j-driver": "^4.4.7"
